Increasing to maximum possible strictness for both C and C++ considering limitations for the PnP C SDK for C and OSConfig Platform tests for C++

+++ src/common/commonutils/CommandUtils.c
@@ -295,7 +295,7 @@ int ExecuteCommand(void* context, const char* command, bool replaceEol, bool for
             if (fileSize > 0)
             {
                 // Truncate to desired maximum, if any
-                if ((maxTextResultBytes > 0) && ((fileSize + 1) > maxTextResultBytes))
+                if ((maxTextResultBytes > 0) && (((size_t)fileSize + 1) > maxTextResultBytes))
                 {
                     fileSize = (maxTextResultBytes > 1) ? (maxTextResultBytes - 1) : 0;
                 }
